+++ packages/terra-clinical-action-header/CHANGELOG.md
@@ -3,6 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* webdriver.io tests for `terra-clinical-action-header`
+
+### Removed
+* Nightwatch tests in favor of webdriver.io tests
 
 2.0.0 - (February 19, 2018)
 ----------